xantispam: handle busy and afk situations gracefully
This means that when the user sets themselves to Busy, by default no
queries are generated and decisions are made by rules.  When the user
is AFK, they can optionally disable the generation of queries with a
"&-ConfigNoQueriesWhenAFK" rule.  They also can keep queries enabled
while Busy with a "&-ConfigQueriesWhenBUSY" rule.