Forget the last commit, this one undoes it:
added setting RtyInvertIgnoreGroupChat:

This reverses the meaning of the checkboxes in the group list in the
communications floater, i. e. when the chat for a group is enabled in
the group list and RtyInvertIgnoreGroupChat is enabled, the group chat
is disabled.

The purpose is to be able to disable and to enable all group chat with
a single setting instead of having to go through all groups to switch,
while being able to except a few groups.

An xantispam rule might be added in the future to specify the
exceptions in such a way that chat for groups excepted by the rule is
always enabled, no matter whether the checkbox is checked or not.