autologin ...
	  ... finally works!

At first I used "--autologin Ratany", which is probably incorrect, yet
didn't give me the usual message about not being able to connect to a
simulator.  Now "--autologin --slurl 'secondlife://blahfoo ..." works
and gets me logged in at the place I want to log in.

The modified timeout is probably not needed at all, but I'm leaving it
because it's working now.