trying to move the "Stand Up" botton out of the way of the communications floater
That moves the other buttons as well, and they might be off-screen
when you're using a small window, so it isn't perfect.  This has
annoyed me for years: Who was so smart to put these buttons right over
the communications floater and to make it so that you cannot move
them?