now spam-filtering message content done right:
You can specify regular expressions with xantispam rules now.  Each
rule comes with an integer which is the score to add to a total for
the message which is being checked.  The score of each matching rule
adds to the total score until the total score goes over a threshold.
When the total score is over the threshold, the message is considered
spam.

The message is replaced with "[blocked spam from]", and the sender
gets a message "[spam message blocked by recipients client]".  You'll
be notified about this by getting an IM sent to yourself.  The
original message as was sent will be logged unaltered if you have
logging enabled.

The 'from' part of the xantispam rule gives the regular expression.
For example:

| come.*http | &-RxScore!6

Since the pipe sign is used as a column seperator, you currently
cannot use it to specify regular expressions.  I'll have to see what I
can do about that.  In the meantime, just use multiple rules when you
need an OR.

Same goes for the colon.  Since it's used as a wildcard, you currently
cannot use it within regular expressions.

Two new settings accompany this:

AntiSpamXtendedRxMinLength

	The default is 85 characters (UTF-8).  When a message is
	shorter, it won't be checked.

AntiSpamXtendedRxScoreThreshold

	This is the threshold mentioned above.  The default is 5.

You'll find those among the debug settings.

As to the regular expression syntax:  The boost library provides
'regex_search()', which is used for the matching, so you can always
look it up if in doubt (seems pretty much like what perl uses).

If you have not specified any '&-RxScore!' rules, processing will not
find any matches, and no messages will be classified as spam.  You
might want to set the threshold very high in that case because that
saves the lookups of the rules.

Only the rules currently in the cache (i. e. in memory) are being
checked.  This is unlikely to change because resorting to the rules
stored in the blacklist seems a bit unreasonable here
performance-wise.  I'm still a long way away from overflowing the rule
caches (wich hold up to 64k rules).

Be careful not to omit the score from a '&-RxScore!' rule, and
especially don't put anything but an integer after the exclamation
mark.  When the rule is too short, i. e. doesn't have anything after
the exclamation mark, it will be ignored.  Put something else than an
integer, and unexpected things might happen.